Update README.md
Tarek found this trick and it worked for both him and me, so I thought this should be in the main wiki, as it's a problem affecting several people.
Running katkin build without -j2 would cause the PC to freeze for eternity. -j2 should limit the amount of cores used and allow the system to remain functional whilst building.